move ConfigFileTraversal from querydb to semanticcpg
